+++ yash-builtin/CHANGELOG.md
@@ -11,12 +11,30 @@ A _private dependency_ is used internally and not visible to downstream users.
 
 ## [0.18.0] - Unreleased
 
+### Added
+
+- The `wait` built-in now interrupts the interactive shell when `SIGINT` is
+  caught with the default action while waiting for jobs.
+
 ### Changed
 
+- The `eval`, `.`/`source`, `fg`, and `wait` built-ins returned from `iter()`
+  now have `handles_signals_internally` set to `true`, indicating that these
+  built-ins handle signals themselves and the caller must not perform additional
+  signal checks while they are executing.
 - Public dependency versions:
     - Rust 1.87.0 → 1.96.0
     - yash-env 0.14.0 → 0.15.0
     - yash-semantics (optional) 0.16.0 → 0.17.0
+
+### Fixed
+
+- Previously, the `fg` built-in always returned an interrupt when invoked in an
+  interactive shell (unless there was an error). This was undesirable because
+  any commands that followed `fg` would unconditionally be ignored. Now, `fg`
+  only returns an interrupt in the following cases:
+    - The job was suspended, or
+    - The job was terminated by `SIGINT` and the `SIGINT` trap is not set.
 
 ## [0.17.0] - 2026-05-23