+++ yash-cli/CHANGELOG.md
@@ -9,6 +9,22 @@ used by other programs.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - Unreleased
+
+### Added
+
+- The interactive shell now allows interrupting command execution with the
+  `SIGINT` signal.
+
+### Fixed
+
+- Previously, the `fg` built-in always returned an interrupt when invoked in an
+  interactive shell (unless there was an error). This was undesirable because
+  any commands that followed `fg` would unconditionally be ignored. Now, `fg`
+  only returns an interrupt in the following cases:
+    - The job was suspended, or
+    - The job was terminated by `SIGINT` and the `SIGINT` trap is not set.
+
 ## [3.0.8] - 2026-05-23
 
 ### Fixed
@@ -326,6 +342,7 @@ later.
 
 - Initial release of the shell
 
+[3.1.0]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.1.0
 [3.0.8]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.8
 [3.0.7]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.7
 [3.0.5]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.5